+++ next.config.mjs
@@ -41,24 +41,6 @@ const nextConfig = {
         ],
       },
       {
-        // * Service Worker specific headers
-        source: '/sw.js',
-        headers: [
-          {
-            key: 'Content-Type',
-            value: 'application/javascript; charset=utf-8',
-          },
-          {
-            key: 'Cache-Control',
-            value: 'no-cache, no-store, must-revalidate',
-          },
-          {
-            key: 'Content-Security-Policy',
-            value: "default-src 'self'; script-src 'self' 'unsafe-inline'; style-src 'self' 'unsafe-inline';",
-          },
-        ],
-      },
-      {
         // * Manifest specific headers
         source: '/manifest.json',
         headers: [